railway-api/requirements.txt
# --- Core API & Web ---
fastapi
uvicorn
requests
httpx
python-dotenv

# --- AI & ML ---
openai
google-generativeai
groq
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib
numpy
pandas

# --- YouTube & Data ---
yt-dlp>=2025.01.01
google-api-python-client
youtube-transcript-api
pytrends
supabase>=2.3.0
psycopg2-binary>=2.9.9

# --- Media Processing (Optional/Premium) ---
openai-whisper
faster-whisper
pillow>=10.0.0
opencv-python-headless>=4.8.0

# --- Integrations ---
stripe